simulator/requirements.txt
rich>=13.0.0
textual>=0.40.0
watchdog>=3.0.0
python-telegram-bot>=21.0
aiohttp>=3.9.0
